+++ .github/workflows/scripts/install-tools.sh
@@ -1,2 +1,2 @@
-dart pub global activate melos 2.9.0
+dart pub global activate melos
 melos bootstrap