+++ flutter_local_notifications/test/macos_flutter_local_notifications_test.dart
@@ -217,6 +217,7 @@ void main() {
               'notification body',
               repeatInterval,
               notificationDetails,
+              androidScheduleMode: AndroidScheduleMode.exact,
             );
 
             expect(