+++ flutter_local_notifications/android/src/main/java/com/dexterous/flutterlocalnotifications/models/NotificationDetails.java
@@ -30,6 +30,7 @@ public class NotificationDetails implements Serializable {
   private static final String MILLISECONDS_SINCE_EPOCH = "millisecondsSinceEpoch";
   private static final String CALLED_AT = "calledAt";
   private static final String REPEAT_INTERVAL = "repeatInterval";
+  private static final String REPEAT_INTERVAL_MILLISECONDS = "repeatIntervalMilliseconds";
   private static final String REPEAT_TIME = "repeatTime";
   private static final String PLATFORM_SPECIFICS = "platformSpecifics";
   private static final String AUTO_CANCEL = "autoCancel";
@@ -145,6 +146,7 @@ public class NotificationDetails implements Serializable {
   public NotificationStyle style;
   public StyleInformation styleInformation;
   public RepeatInterval repeatInterval;
+  public Integer repeatIntervalMilliseconds;
   public Time repeatTime;
   public Long millisecondsSinceEpoch;
   public Long calledAt;
@@ -225,6 +227,10 @@ public class NotificationDetails implements Serializable {
     if (arguments.containsKey(REPEAT_INTERVAL)) {
       notificationDetails.repeatInterval =
           RepeatInterval.values()[(Integer) arguments.get(REPEAT_INTERVAL)];
+    }
+    if (arguments.containsKey(REPEAT_INTERVAL_MILLISECONDS)) {
+      notificationDetails.repeatIntervalMilliseconds =
+          (Integer) arguments.get(REPEAT_INTERVAL_MILLISECONDS);
     }
     if (arguments.containsKey(REPEAT_TIME)) {
       @SuppressWarnings("unchecked")
@@ -458,7 +464,7 @@ public class NotificationDetails implements Serializable {
         result.add(
             new MessageDetails(
                 (String) messageData.get(TEXT),
-                (Long) messageData.get(TIMESTAMP),
+                LongUtils.parseLong(messageData.get(TIMESTAMP)),
                 readPersonDetails((Map<String, Object>) messageData.get(PERSON)),
                 (String) messageData.get(DATA_MIME_TYPE),
                 (String) messageData.get(DATA_URI)));