+++ flutter_local_notifications/lib/src/flutter_local_notifications_plugin.dart
@@ -38,9 +38,6 @@ class FlutterLocalNotificationsPlugin {
     } else if (defaultTargetPlatform == TargetPlatform.macOS) {
       FlutterLocalNotificationsPlatform.instance =
           MacOSFlutterLocalNotificationsPlugin();
-    } else if (defaultTargetPlatform == TargetPlatform.linux) {
-      FlutterLocalNotificationsPlatform.instance =
-          LinuxFlutterLocalNotificationsPlugin();
     }
   }
 
@@ -301,13 +298,6 @@ class FlutterLocalNotificationsPlugin {
   /// platform channel in yyyy-mm-dd hh:mm:ss format. Therefore, the precision
   /// is at the best to the second.
   ///
-  /// The [androidAllowWhileIdle] parameter determines if the notification
-  /// should still be shown at the exact time even when the device is in a
-  /// low-power idle mode. This parameter has been deprecated and will removed
-  /// in a future major release in favour of the [androidScheduledMode]
-  /// parameter that provides the same functionality in addition to being able
-  /// to schedule notifications with inexact timings.
-  ///
   /// The [uiLocalNotificationDateInterpretation] is for iOS versions older
   /// than 10 as the APIs have limited support for time zones. With this
   /// parameter, it is used to determine if the scheduled date should be
@@ -339,9 +329,7 @@ class FlutterLocalNotificationsPlugin {
     NotificationDetails notificationDetails, {
     required UILocalNotificationDateInterpretation
         uiLocalNotificationDateInterpretation,
-    @Deprecated('Deprecated in favor of the androidScheduleMode parameter')
-    bool androidAllowWhileIdle = false,
-    AndroidScheduleMode? androidScheduleMode,
+    required AndroidScheduleMode androidScheduleMode,
     String? payload,
     DateTimeComponents? matchDateTimeComponents,
   }) async {
@@ -352,14 +340,9 @@ class FlutterLocalNotificationsPlugin {
       await resolvePlatformSpecificImplementation<
               AndroidFlutterLocalNotificationsPlugin>()!
           .zonedSchedule(
-              id,
-              title,
-              body,
-              scheduledDate,
-              notificationDetails.android,
+              id, title, body, scheduledDate, notificationDetails.android,
               payload: payload,
-              scheduleMode: _chooseScheduleMode(
-                  androidScheduleMode, androidAllowWhileIdle),
+              scheduleMode: androidScheduleMode,
               matchDateTimeComponents: matchDateTimeComponents);
     } else if (defaultTargetPlatform == TargetPlatform.iOS) {
       await resolvePlatformSpecificImplementation<
@@ -388,18 +371,6 @@ class FlutterLocalNotificationsPlugin {
   /// notification will be an hour after the method has been called and
   /// then every hour after that.
   ///
-  /// If [androidAllowWhileIdle] is `false`, the Android `AlarmManager` APIs
-  /// are used to set a recurring inexact alarm that would present the
-  /// notification. This means that there may be delay in on when
-  /// notifications are displayed. If [androidAllowWhileIdle] is `true`, the
-  /// Android `AlarmManager` APIs are used to schedule a single notification
-  /// to be shown at the exact time even when the device is in a low-power idle
-  /// mode. After it is shown, the next one would be scheduled and this would
-  /// repeat. Note that this parameter has been deprecated and will removed in
-  /// future majorrelease in favour of the [androidScheduledMode] parameter that
-  /// provides the same functionality in addition to being able to schedule
-  /// notifications with inexact timings.
-  ///
   /// On Android, this will also require additional setup for the app,
   /// especially in the app's `AndroidManifest.xml` file. Please see check the
   /// readme for further details.
@@ -409,10 +380,8 @@ class FlutterLocalNotificationsPlugin {
     String? body,
     RepeatInterval repeatInterval,
     NotificationDetails notificationDetails, {
+    required AndroidScheduleMode androidScheduleMode,
     String? payload,
-    @Deprecated('Deprecated in favor of the androidScheduleMode parameter')
-    bool androidAllowWhileIdle = false,
-    AndroidScheduleMode? androidScheduleMode,
   }) async {
     if (kIsWeb) {
       return;
@@ -423,8 +392,7 @@ class FlutterLocalNotificationsPlugin {
           ?.periodicallyShow(id, title, body, repeatInterval,
               notificationDetails: notificationDetails.android,
               payload: payload,
-              scheduleMode: _chooseScheduleMode(
-                  androidScheduleMode, androidAllowWhileIdle));
+              scheduleMode: androidScheduleMode);
     } else if (defaultTargetPlatform == TargetPlatform.iOS) {
       await resolvePlatformSpecificImplementation<
               IOSFlutterLocalNotificationsPlugin>()
@@ -441,12 +409,54 @@ class FlutterLocalNotificationsPlugin {
     }
   }
 
-  AndroidScheduleMode _chooseScheduleMode(
-          AndroidScheduleMode? scheduleMode, bool allowWhileIdle) =>
-      scheduleMode ??
-      (allowWhileIdle
-          ? AndroidScheduleMode.exactAllowWhileIdle
-          : AndroidScheduleMode.exact);
+  /// Periodically show a notification using the specified custom duration
+  /// interval.
+  ///
+  /// For example, specifying a 5 minutes repeat duration interval means
+  /// the first time the notification will be an 5 minutes after the method
+  /// has been called and then every 5 minutes after that.
+  ///
+  /// On Android, this will also require additional setup for the app,
+  /// especially in the app's `AndroidManifest.xml` file. Please see check the
+  /// readme for further details.
+  Future<void> periodicallyShowWithDuration(
+    int id,
+    String? title,
+    String? body,
+    Duration repeatDurationInterval,
+    NotificationDetails notificationDetails, {
+    AndroidScheduleMode androidScheduleMode = AndroidScheduleMode.exact,
+    String? payload,
+  }) async {
+    if (kIsWeb) {
+      return;
+    }
+    if (defaultTargetPlatform == TargetPlatform.android) {
+      await resolvePlatformSpecificImplementation<
+              AndroidFlutterLocalNotificationsPlugin>()
+          ?.periodicallyShowWithDuration(
+              id, title, body, repeatDurationInterval,
+              notificationDetails: notificationDetails.android,
+              payload: payload,
+              scheduleMode: androidScheduleMode);
+    } else if (defaultTargetPlatform == TargetPlatform.iOS) {
+      await resolvePlatformSpecificImplementation<
+              IOSFlutterLocalNotificationsPlugin>()
+          ?.periodicallyShowWithDuration(
+              id, title, body, repeatDurationInterval,
+              notificationDetails: notificationDetails.iOS, payload: payload);
+    } else if (defaultTargetPlatform == TargetPlatform.macOS) {
+      await resolvePlatformSpecificImplementation<
+              MacOSFlutterLocalNotificationsPlugin>()
+          ?.periodicallyShowWithDuration(
+              id, title, body, repeatDurationInterval,
+              notificationDetails: notificationDetails.macOS, payload: payload);
+    } else {
+      await FlutterLocalNotificationsPlatform.instance
+          .periodicallyShowWithDuration(
+              id, title, body, repeatDurationInterval);
+    }
+  }
 
   /// Returns a list of notifications pending to be delivered/shown.
   Future<List<PendingNotificationRequest>> pendingNotificationRequests() =>