+++ flutter_local_notifications_platform_interface/CHANGELOG.md
@@ -1,3 +1,6 @@
+## [vNext]
+
+* **Breaking change** Bumped minimum Flutter SDK requirement to 3.13
 
 ## [7.2.0]